ml/ml-scikit-learn/python-model-offline/src/main/resources/requirements.txt
pandas
sklearn
datetime
flask
flask_restful
pyyaml
